// client/src/shared/api/models/PrizeOutAddUUID.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { PrizeType } from './PrizeType';
export type PrizeOutAddUUID = {
    prize_uuid: string;
    prize_type: PrizeType;
    expires_at?: (string | null);
};