// client/src/shared/api/models/PrizeType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum PrizeType {
    _3_DISCOUNT = '3%_discount',
    _5_DISCOUNT = '5%_discount',
    _7_INCREASED_REFERRAL_BONUS = '7%_increased_referral_bonus',
    _03_DEPOSIT = '03_deposit',
}